Manejo excepciones Genre

+++ src/main/java/com/lucascordoba/backendchallenge/services/GenreServiceImpl.java
@@ -1,25 +1,26 @@
 package com.lucascordoba.backendchallenge.services;
 
 import com.lucascordoba.backendchallenge.dto.GenreDTO;
-import com.lucascordoba.backendchallenge.models.CharacterModel;
 import com.lucascordoba.backendchallenge.models.GenreModel;
 import com.lucascordoba.backendchallenge.repositories.GenreRepository;
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 @Service
-public class GenreServiceImpl implements GenreService{
+public class GenreServiceImpl implements GenreService {
     @Autowired
     GenreRepository genreRepository;
+
     @Override
     @Transactional(readOnly = true)
     public List<GenreDTO> listGenres() {
-        List<GenreModel> entities=genreRepository.findAll();
-        List<GenreDTO> dtos= new ArrayList<>();
+        List<GenreModel> entities = genreRepository.findAll();
+        List<GenreDTO> dtos = new ArrayList<>();
         for (GenreModel entity : entities) {
             dtos.add(GenreDTO.simpleFrom(entity));
         }
@@ -29,35 +30,35 @@ public class GenreServiceImpl implements GenreService{
     @Override
     @Transactional(readOnly = true)
     public GenreDTO findGenre(Long id) {
-        GenreModel entity=genreRepository.getById(id);
+        GenreModel entity = genreRepository.getById(id);
         return GenreDTO.from(entity);
     }
 
     @Override
     @Transactional
     public GenreDTO insertGenre(GenreDTO genreDTO) {
-        GenreModel entity=genreDTO.buildEntity();
+        Optional<GenreModel> genreOptional = genreRepository.findByName(genreDTO.getName());
+        if (genreOptional.isPresent()) {
+            throw new IllegalStateException("duplicated genre name");
+        }
+        GenreModel entity = genreDTO.buildEntity();
         return GenreDTO.from(genreRepository.save(entity));
     }
 
     @Override
     @Transactional
-    public Boolean deleteGenre(GenreDTO genreDTO) {
-        try {
+    public void deleteGenre(GenreDTO genreDTO) {
+        if (!genreRepository.existsById(genreDTO.getId()))
+            throw new IllegalStateException("Inexistent genre with id: " + genreDTO.getId());
+        else
             genreRepository.deleteById(genreDTO.getId());
-            return true;
-        }catch (Exception e){
-            return false;
-        }
     }
-    public Boolean deleteGenre(Long id) {
-        try {
+    @Override
+    @Transactional
+    public void deleteGenre(Long id) {
+        if (!genreRepository.existsById(id))
+            throw new IllegalStateException("Inexistent genre with id: " + id);
+        else
             genreRepository.deleteById(id);
-            return true;
-        }catch (Exception e){
-            return false;
-        }
     }
-
-
 }
